Refactor: sep file
